Moved remux function to util.extension, fixed exception handling, added docs
